refactor: zustand store 변경된 이름 경로 변경
